Some minor fixes to ColumnChunkPageStore and friends.

+++ DB/src/main/java/io/deephaven/db/v2/locations/parquet/FixedPageSizeColumnChunkPageStore.java
@@ -58,7 +58,7 @@ class FixedPageSizeColumnChunkPageStore<ATTR extends Attributes.Any> extends Col
 
                 if (page == null) {
                     try {
-                        page = new IntrusivePage<>(toPage(pageNum * pageFixedSize, columnPageReaders[pageNum]));
+                        page = new IntrusivePage<>(toPage((long) pageNum * pageFixedSize, columnPageReaders[pageNum]));
                     } catch (IOException except) {
                         throw new UncheckedIOException(except);
                     }
